鳥が放たれた数の０のチラ見えを無くした
